chore: cleanup this repo

+++ src/tables/evals/BetStatistics.java
@@ -4,7 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * Full information on a bet, along with information for each payline.
+ * Aggregated statistics for a single bet, containing individual statistics
+ * for each payline defined in the bet.
+ * <p>
+ * This class accumulates combination counts during enumeration and finalizes
+ * hit frequency, return percentage, and house edge calculations once all
+ * combinations have been tallied.
  */
 public class BetStatistics {
 
@@ -14,11 +19,27 @@ public class BetStatistics {
     private double hitFrequency;
     private double pctReturn;
 
+    /**
+     * Constructs bet statistics for the given bet name and paylines.
+     *
+     * @param name  the name of the bet
+     * @param stats the statistics trackers for each payline
+     */
     public BetStatistics(String name, PaylineStatistics... stats) {
         this.name = name;
         this.stats = List.of(stats);
     }
 
+    /**
+     * Returns the first payline statistics whose hit evaluator matches the
+     * given hands, or {@code null} if no payline matches.
+     * <p>
+     * Paylines are checked in order, so the ordering defined in the
+     * {@link BetEvaluator} determines priority.
+     *
+     * @param hands the hands to evaluate
+     * @return the matching payline statistics, or null
+     */
     public PaylineStatistics getHitStat(Hand... hands) {
         for (PaylineStatistics stat : stats) {
             if (stat.getPayEvaluator().hitEvaluator().isHit(hands)) {
@@ -28,10 +49,21 @@ public class BetStatistics {
         return null;
     }
 
+    /**
+     * Adds the given combination count to all payline statistics.
+     *
+     * @param combos the number of combinations to add
+     */
     public void addCombos(long combos) {
         stats.forEach(stat -> stat.addCombinations(combos));
     }
 
+    /**
+     * Finalizes all statistics by computing hit frequencies and return percentages
+     * relative to the total number of possible combinations.
+     *
+     * @param totalCombos the total number of all possible hand combinations
+     */
     public void setTotalCombos(long totalCombos) {
         this.totalCombos = totalCombos;
         for (PaylineStatistics stat : stats) {
@@ -41,18 +73,22 @@ public class BetStatistics {
         }
     }
 
+    /** Returns the list of per-payline statistics. */
     public List<PaylineStatistics> getStats() {
         return stats;
     }
 
-    public double getTotalCombos() {
+    /** Returns the total number of combinations evaluated. */
+    public long getTotalCombos() {
         return totalCombos;
     }
 
+    /** Returns the overall hit frequency across all paylines. */
     public double getHitFrequency() {
         return hitFrequency;
     }
 
+    /** Returns the overall expected return (as a decimal, e.g., 0.986 for 98.6%). */
     public double getPctReturn() {
         return pctReturn;
     }
@@ -68,10 +104,10 @@ public class BetStatistics {
                 "\nPct Return\t" +
                 pctReturn +
                 "\nHouse Edge\t" +
-                (1-pctReturn) + "\n\n" +
+                (1 - pctReturn) + "\n\n" +
                 "Bet\tCombos\tHit Frequency\tPct Return\n" +
-                stats.stream().map(s ->
-                            s.getPayEvaluator().name() +
+                stats.stream()
+                        .map(s -> s.getPayEvaluator().name() +
                                 "\t" +
                                 s.getCombos() +
                                 "\t" +
@@ -81,3 +117,4 @@ public class BetStatistics {
                         .collect(Collectors.joining("\n"));
     }
 }
+
